+++ test/integration/full/target-size/too-many-rects.js
@@ -18,7 +18,6 @@ describe('target-size too many rects test', () => {
           done(err);
         }
         results = r;
-        console.log(results);
         done();
       });
     });